新增子订单表

+++ booking-internal-service/src/main/java/com/plateno/booking/internal/service/order/OrderInsertActorService.java
@@ -114,7 +114,7 @@ public class OrderInsertActorService {
             conditions.setOrderAmount(orderCheckDetail.getCouponOrderAmount());
             if(CollectionUtils.isNotEmpty(orderCheckDetail.getCouponProductList())){
                 conditions.setProductId(orderCheckDetail.getCouponProductList().get(0).getSpuId().intValue());
-                conditions.setCategoryId(orderCheckDetail.getCouponProductList().get(0).getCategoryId().intValue());
+                conditions.setCategoryId(orderCheckDetail.getCouponProductList().get(0).getCategoryId());
             }
             ResultVo<BaseResponse> useCouponResult = couponService.useCoupon(useCouponParam);
             if (!useCouponResult.success()) {
